+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 <!-- Add new lines here. Version number will be decided later -->
 
 - ...
+- feat: Tweak rules for async, destructuring and TypeScript
 
 ## 1.0.0